pull out pacer and distribute funds
